Moving scrapers to a separate folder.
