+++ scrapers/anime_list_scraper.rb
