+++ scrapers/anime_scraper.rb
