+++ scrapers/anime_search_scraper.rb
