+++ scrapers/manga_list_scraper.rb
