+++ scrapers/manga_scraper.rb
