+++ scrapers/manga_search_scraper.rb
